+++ life_expectancy/cleaning.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 import pandas as pd
