+++ life_expectancy/tests/conftest.py
@@ -17,13 +17,13 @@ def run_before_and_after_tests() -> None:
     file_path.unlink(missing_ok=True)
 
 
+@pytest.fixture(scope="module")
+def eu_life_expectancy_raw() -> pd.DataFrame:
+    """Fixture to load the expected output of the cleaning script"""
+    return pd.read_csv(FIXTURES_DIR / "eu_life_expectancy_raw.tsv", sep="\t")
+
 @pytest.fixture(scope="session")
 def pt_life_expectancy_expected() -> pd.DataFrame:
     """Fixture to load the expected output of the cleaning script"""
     return pd.read_csv(FIXTURES_DIR / "pt_life_expectancy_expected.csv")
 
-
-@pytest.fixture(scope="session")
-def eu_life_expectancy_raw() -> pd.DataFrame:
-    """Fixture to load the expected output of the cleaning script"""
-    return pd.read_csv(OUTPUT_DIR / "eu_life_expectancy_raw.tsv", sep='\t')
